refactor: cải thiện kiến trúc và sửa lỗi hệ thống

+++ src/main/java/minesweeper/model/User.java
@@ -3,7 +3,7 @@ package minesweeper.model;
 import java.time.LocalDateTime;
 
 public class User {
-    private int id;
+    private long id;
     private String username;
     private String passwordHash; // BCrypt hash
     private String displayName;
@@ -28,7 +28,7 @@ public class User {
     }
 
     // ── Getters ────────────────────────────────────────
-    public int getId() { return id; }
+    public long getId() { return id; }
     public String getUsername() { return username; }
     public String getPasswordHash() { return passwordHash; }
     public String getDisplayName() { return displayName; }
@@ -39,7 +39,7 @@ public class User {
     public LocalDateTime getLastLoginAt() { return lastLoginAt; }
 
     // ── Setters ────────────────────────────────────────
-    public void setId(int id) { this.id = id; }
+    public void setId(long id) { this.id = id; }
     public void setUsername(String username) { this.username = username; }
     public void setPasswordHash(String passwordHash) { this.passwordHash = passwordHash; }
     public void setDisplayName(String displayName) { this.displayName = displayName; }
